kafka_streaming/requirements.txt
kafka-python
scikit-learn
pandas
joblib
numpy
